+++ CHANGELOG.md
@@ -204,3 +204,10 @@ Please add a _short_ line describing the PR you make, if the PR implements a spe
 
 - Limit projects listing to active projects only; a `--show-all` flag can be used for listing all projects, active and inactive ([#556](https://github.com/ScilifelabDataCentre/dds_cli/pull/556))
 - Display name of creator when listing projects ([#557](https://github.com/ScilifelabDataCentre/dds_cli/pull/557))
+- New command: `dds project info [OPTIONS]` to display information about specific project ([#561](https://github.com/ScilifelabDataCentre/dds_cli/pull/561))
+- New option for `dds data put`: `--destination` - allow upload to existing or new directory ([#559](https://github.com/ScilifelabDataCentre/dds_cli/pull/559))
+
+## Sprint (2022-10-28 - 2022-11-11)
+
+- Bug: Catch Timeout exception for when requests are too slow ([#563](https://github.com/ScilifelabDataCentre/dds_cli/pull/563))
+- Check that `dds data get` is used with either `--get-all` or `--source`/`--source-path-fail` and display appropriate message ([#564](https://github.com/ScilifelabDataCentre/dds_cli/pull/564))